Update VIPER module
