--- DemoViper/Sources/Modules/Detail/Interactor/RepoInteractorInput.swift
@@ -1,16 +1,0 @@
-//
-//  RepoInteractorInput.swift
-//  DemoViper
-//
-//  Created by Su Van Ho on 13/8/19.
-//  Copyright © 2019 Nimble Co. Ltd. All rights reserved.
-//
-
-import Foundation
-
-// sourcery: AutoMockable
-protocol RepoInteractorInput: AnyObject {
-    func updateRepoId(_ id: Int)
-    func getInfo()
-    func authorPath() -> String
-}
