--- DemoViper/Sources/Modules/Detail/Presenter/RepoInput.swift
@@ -1,14 +1,0 @@
-//
-//  RepoInput.swift
-//  DemoViper
-//
-//  Created by Su Van Ho on 13/8/19.
-//  Copyright © 2019 Nimble Co. Ltd. All rights reserved.
-//
-
-import Foundation
-
-// sourcery: AutoMockable
-protocol RepoInput: AnyObject {
-    func apply(repoId: Int)
-}
