+++ DemoViper/Sources/Modules/Detail/View/RepoViewItem.swift
@@ -13,4 +13,5 @@ struct RepoItem {
     let repoFullName: String
     let description: String
     let authorName: String
+    let didView: Bool
 }
