--- DemoViper/Sources/Modules/Detail/View/RepoViewOutput.swift
@@ -1,15 +1,0 @@
-//
-//  RepoViewOutput.swift
-//  DemoViper
-//
-//  Created by Su Van Ho on 13/8/19.
-//  Copyright © 2019 Nimble Co. Ltd. All rights reserved.
-//
-
-import Foundation
-
-// sourcery: AutoMockable
-protocol RepoViewOutput: AnyObject {
-    func viewDidLoad()
-    func openWebPage()
-}
