--- DemoViper/Sources/Modules/ListRepos/Interactor/ReposInteractorInput.swift
@@ -1,15 +1,0 @@
-//
-//  ReposInteractorInput.swift
-//  DemoViper
-//
-//  Created by Su Van Ho on 13/8/19.
-//  Copyright © 2019 Nimble Co. Ltd. All rights reserved.
-//
-
-import Foundation
-
-// sourcery: AutoMockable
-protocol ReposInteractorInput: AnyObject {
-    func getRepos()
-    func getRepoId(at index: Int) -> Int?
-}
