--- DemoViper/Sources/Modules/ListRepos/Presenter/ReposOutput.swift
@@ -1,14 +1,0 @@
-//
-//  ReposOutput.swift
-//  DemoViper
-//
-//  Created by Su Van Ho on 13/8/19.
-//  Copyright © 2019 Nimble Co. Ltd. All rights reserved.
-//
-
-import Foundation
-
-// sourcery: AutoMockable
-protocol ReposOutput: AnyObject {
-
-}
