+++ DemoViper/Sources/Modules/ListRepos/Router/ReposRouter.swift
@@ -8,6 +8,12 @@
 
 import UIKit
 
+// sourcery: AutoMockable
+protocol ReposRouterInput: AnyObject {
+    func detail(with id: Int)
+    func showError(_ error: Error)
+}
+
 final class ReposRouter {
     
     weak var view: ReposViewInput?
