--- DemoViper/Sources/Modules/ListRepos/View/ReposViewInput.swift
@@ -1,16 +1,0 @@
-//
-//  ReposViewInput.swift
-//  DemoViper
-//
-//  Created by Su Van Ho on 13/8/19.
-//  Copyright © 2019 Nimble Co. Ltd. All rights reserved.
-//
-
-import Foundation
-
-// sourcery: AutoMockable
-protocol ReposViewInput: AnyObject {
-    func configure()
-    func showEmptyMessage()
-    func showData(_ data: [String])
-}
